@@ -1,17 +1,16 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from "react-dom/client";
 
-import Header from './Header.jsx';
+import Header from "./Header.jsx";
 
 const App = () => {
   return (
-    <div className='container'>
-      <Header/>
+    <div className="container">
+      <Header />
       <div className="container-view">Some activities should be here</div>
     </div>
   );
 };
 
-ReactDOM.render(<App/>, document.getElementById('app'));
-
+const root = createRoot(document.getElementById("app"));
+root.render(<App tab="home" />);
 export default App;
